refactor: split repository-path vs solution-path tests
